refactor: KanbanLayout into smaller components with hooks
Breaks down monolithic KanbanLayout component into smaller, more manageable pieces:

- Extracts core functionality into custom hooks (useAccordion, useDragAndDrop, useFilters, useKanbanData)
- Creates separate component files for better organization and maintainability
- Moves layout-specific code to new KanbanLayout component
- Improves code reusability and separation of concerns

This refactoring makes the codebase more maintainable and easier to test by following single responsibility principle and keeping components focused.
