+++ src/app/index.tsx
@@ -1,9 +1,5 @@
-import { KanbanLayout } from '@/features/kanban/KanbanLayout';
+import { KanbanLayout } from '@/features/kanban/components/KanbanLayout';
 
 export function App() {
-  return (
-    <>
-      <KanbanLayout />
-    </>
-  );
+  return <KanbanLayout />;
 }
